+++ src/OpenTelemetry.Instrumentation.AspNetCore/AspNetCoreInstrumentationOptions.cs
@@ -40,7 +40,7 @@ namespace OpenTelemetry.Instrumentation.AspNetCore
         /// <para><see cref="Activity"/>: the activity being enriched.</para>
         /// <para><see cref="HttpRequest"/>: the HttpRequest object from which additional information can be extracted to enrich the activity during request start event.</para>
         /// </remarks>
-        public Action<Activity, HttpRequest> EnrichOnStart { get; set; }
+        public Action<Activity, HttpRequest> EnrichWithHttpRequest { get; set; }
 
         /// <summary>
         /// Gets or sets an action to enrich an Activity during stop event.
@@ -49,7 +49,7 @@ namespace OpenTelemetry.Instrumentation.AspNetCore
         /// <para><see cref="Activity"/>: the activity being enriched.</para>
         /// <para><see cref="HttpResponse"/>: the HttpResponse object from which additional information can be extracted to enrich the activity during request stop event.</para>
         /// </remarks>
-        public Action<Activity, HttpResponse> EnrichOnStop { get; set; }
+        public Action<Activity, HttpResponse> EnrichWithHttpResponse { get; set; }
 
         /// <summary>
         /// Gets or sets an action to enrich an Activity during exception event.
@@ -58,7 +58,7 @@ namespace OpenTelemetry.Instrumentation.AspNetCore
         /// <para><see cref="Activity"/>: the activity being enriched.</para>
         /// <para><see cref="Exception"/>: the Exception object from which additional information can be extracted to enrich the activity on exception during request.</para>
         /// </remarks>
-        public Action<Activity, Exception> EnrichOnException { get; set; }
+        public Action<Activity, Exception> EnrichWithException { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating whether the exception will be recorded as ActivityEvent or not.