+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* `Enrich` callback is now split into 3 different callbacks based on the object
+that is passed in to enrich activity. The new callbacks are
+`EnrichWithHttpRequest`, `EnrichWithHttpResponse` and `EnrichWithException`.
+([#3749](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3749))
+
 ## 1.0.0-rc9.8
 
 Released 2022-Oct-17